Tests/LibWeb/Text/expected/HTML/error-event.txt

type: error
message: well hello friends 🐞
filename: file:///tmp/🐞.js
lineno: 12
colno: 34
error identity: true
